query sorting framework for specifying a sort using a declarative name
This allows passing around an enum to specify a sort. This improves on a problem
I've encountered a number of times where specifying how to sort involves passing strings
around. Whether that is passing the name of the column or some other string, which is then
used to apply an OrderBy to an IEnumerable or IQueryable.